fix batch_norm bias_grad loss in cinn
